pyproject.toml
[project]
name = "agent-service-toolkit"
version = "0.1.0"
description = "Full toolkit for running an AI agent service built with LangGraph, FastAPI and Streamlit"
readme = "README.md"
authors = [
    {name = "Joshua Carroll", email = "carroll.joshk@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.12",
]

requires-python = ">=3.9, <=3.12.3"

# NOTE: FastAPI < 0.100.0 and Pydantic v1 is required until langchain has full pydantic v2 compatibility
# https://python.langchain.com/v0.1/docs/guides/development/pydantic_compatibility/
# https://github.com/langchain-ai/langchain/discussions/24405
# https://github.com/langchain-ai/langchain/discussions/9337
# IMPORTANT: This also requires using python < 3.12.4
dependencies = [
    "duckduckgo-search ~=6.2.6",
    "fastapi <0.100.0",
    "httpx~=0.27.2",
    "langchain-core ~=0.2.26",
    "langchain-community ~=0.2.11",
    "langchain-openai ~=0.1.20",
    "langchain-groq ~=0.1.9",
    "langgraph ~=0.2.3",
    "langgraph-checkpoint ~=1.0.2",
    "langgraph-checkpoint-sqlite ~=1.0.0",
    "langsmith ~=0.1.96",
    "numexpr ~=2.10.1",
    "pydantic ~=1.10.17",
    "pyowm ~=3.3.0",
    "python-dotenv ~=1.0.1",
    "setuptools ~=74.0.0",
    "streamlit ~=1.37.0",
    "uvicorn ~=0.30.5",
    "bs4",
    "langchain_huggingface",
    "langchain_experimental",
    "psycopg2-binary ~=2.9.9",
    "dill",
    "spider-client",
    "faiss-cpu",
    "pypdf",
    "python-multipart",
    "googlesearch-python",
    "langchain-ollama"
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
    "pytest-env",
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest_env]
OPENAI_API_KEY = "sk-fake-openai-key"
